Update index.md
I tried to just simplify in most cases. I added some suggestions for `<code>` wrapping property names to call them out, which I think is conventional here.
I changed the example Point coordinates to make it more obvious that it is a latlng number and not something like a Canvas x,y.
Take or leave whatever.