Update index.md
Addressed comments from Cookbook session 22/04: Add a paragraph to the implementation note to mention this recipe uses the navPlace and its a case where the navPlace is located on the Manifest.
Also worth mentioning that navPlace isn’t semantic so not possible to say this is the location of the item.

Added a mention (that can be taken out if not wanted), about an example viewer in the guide, and a clarification in the example about the location shown being its current one.