Removing Annona from as Annotation list viewer for
Annona works if you pass in the Manifest and the Annotation list but it doesn't work if you only pass in the Annotation list. There is no way for Annona to work though as the Annotation list doesn't reference the containing manifest. So its fine to include Annona in the Manifest viewers but not the Annnotation list which can't have any viewers.