Elaborated on dimensional conflicts for viewers ↔️
My initial thought was to comment on how SVG creation tools write their paths, but in conversation with cookbook authors it became clear that there was a bigger bump out there, that of annotations getting created that don't match the dimensions of the resource being annotated (at least as it is represented in a manifest). The latter became the focus of this graf.