Added navDate and navPlace data to Collection 📝
Decision arrived at during cookbook authors call 19 July 2024. Reasoning is partly that we want a viewer to not have to dig into the constituent manifests to get the data and that we want to show how a Collection can be made from manifests that themselves do not have chrono or geo data in them.