Adjusted origin per #631 h/t @stephenwf 🛠️
To get the rotation we want, the point of rotation cannot be the center of a rectangle, given that the implicit overlap between the vertical rectangle of the resource and the horizontal rectangle of the Canvas share a corner rather than their center. Fortunately, figuring the rotation point is, in this case and many others, just figuring the center of the imaginary square formed by the overlap of the resource and Canvas. In turn this center is half the narrower dimension of the resource or Canvas, provided they are the same dimensions